+++ src/entities/routingTable/RouteNode.h
@@ -45,13 +45,13 @@ public:
      * @brief SRTT, smoothed round-trip time (RFC 6298)
      *
      */
-    uint32_t SRTT = 0;
+    unsigned long SRTT = 0;
 
     /**
      * @brief RTTVAR, round-trip time variation (RFC 6298)
      *
      */
-    uint32_t RTTVAR = 0;
+    unsigned long RTTVAR = 0;
 
     /**
      * @brief Construct a new Route Node object
